Crate mwgc

A sesame-seed-sized heap with a tri-color, tracing, conservative, incremental, non-compacting garbage collector, for implementing a tiny language on tiny hardware.

It's simple, not thread safe, and efficient for allocations up to about 512 bytes and heaps under about 1MB.

Here's an example of creating a 256-byte heap on the stack, using it to allocate two different objects, and then running the garbage collector to reap one of them:

use mwgc::Heap;

#[derive(Default)]
struct Toaster { a: u32 }

let mut data: [u8; 256] = [0; 256];
let mut h = Heap::from_bytes(&mut data);
let o1 = h.allocate_object::<Toaster>().unwrap();
h.gc(&[ o1 ]);

Structs

Heap
Takes ownership of a block of Memory, hands out chunks of it, and garbage collects unused chunks on demand.

HeapStats
Stats returned from Heap::get_stats.

Memory
Wrapper for an owned, mutable chunk of memory.