Struct mwgc::Heap

pub struct Heap<'heap> { /* fields omitted */ }

Takes ownership of a block of Memory, hands out chunks of it, and garbage collects unused chunks on demand.

The heap is organized into allocatable blocks, with a small region reserved as a bitmap. The block size is configured by a compile-time constant, using a default of 16 bytes. Each block has a 2-bit "color" in the bitmap, at a cost of about 2% overhead. One "color" is used to mark continuations of contiguous spans, so each allocation can be identified by a color followed by one or more "continue" markers. The other 3 colors correspond to the white, gray, and black colors of tri-color marking, although I've chosen to call them blue, green, and "check".

Separately, a sorted free list is maintained by storing a "next" link and size in each free block. This consumes 8 bytes on a 32-bit system, limiting the minimum block size.

The heap object (its state) should consume about 9 words, or 36 bytes on a 32-bit system.

Methods

impl<'heap> Heap<'heap>

pub fn new(m: Memory<'heap>) -> Heap<'heap>

Create a new heap out of a mutable chunk of memory.

pub fn from_bytes(bytes: &'heap mut [u8]) -> Heap<'heap>

Create a new heap out of a mutable byte-slice.

pub fn allocate(&mut self, amount: usize) -> Option<Memory<'heap>>

Request a amount bytes of memory. The size will be rounded up to a multiple of the block size. Returns None if a block of memory that big isn't available,

pub fn allocate_object<T: Default>(&mut self) -> Option<&'heap mut T>

Request enough memory to hold an object of type T. The object will be initialized to its default value. Returns None if a block of memory that big isn't available.

pub fn allocate_array<T: Default>(
&mut self,
count: usize
) -> Option<&'heap mut [T]>

Request enough memory to hold an array of count objects of type T. Each object in the array will be initialized to its default value. Returns None if a block of memory that big isn't available.

pub fn retire(&mut self, m: Memory<'heap>)

Give back an allocation without waiting for a GC round.

pub fn retire_object<T>(&mut self, obj: &'heap mut T)

Give back an allocated object without waiting for a GC round.

pub fn mark_start<T>(&mut self, roots: &[&T])

Start the first phase of garbage collection. This is only useful if you want tight control over latency -- otherwise, you should call gc().

roots must be a slice of references to objects in the heap which are the roots of your object graph. Successive calls to mark_round() will follow the object graph and mark all live objects. A follow-on call to sweep() will free the remaining spans.

Important: If you modify any objects before mark_round() returns true, you must notify the garbage collector to (re)check the object you modified by calling mark_check.

pub fn mark_round(&mut self) -> bool

Do one "round" of the mark phase of garbage collection. This is only useful if you want tight control over latency -- otherwise, you should call gc().

Walk the current mark span, following links from marked objects to find objects to check on the next round.

Returns true if this phase is over, and all live objects have been marked. You must call sweep() to finish the collection and free up memory.

Important: If you modify any objects before this function returns true, you must notify the garbage collector to (re)check the object you modified by calling mark_check.

pub fn mark<T>(&mut self, roots: &[&T])

Do the mark phase of garbage collection (the first of two phases).

roots must be a slice of references to objects in the heap which are the roots of your object graph. Any object that isn't directly or indirectly referenced by these roots will be freed by a following call to sweep().

This is the equivalent of heap.mark_start(roots); while !heap.mark_round() {};.

pub fn mark_check<T>(&mut self, obj: &T)

Mark an object to be re-checked because it's been modified during the mark phase of garbage collection.

This function is only useful if you use the incremental GC calls mark_start and mark_round.

pub fn get_mark_range(&self) -> (*const u8, *const u8)

For debugging and tests, report the current range of addresses that will be scanned on the next mark_round.

pub fn sweep(&mut self)

Sweep through the heap and move every un-marked span of memory into the free list. This is the 2nd and final phase of garbage collection.

pub fn gc<T>(&mut self, roots: &[&T])

Do an entire GC round, freeing any currently unused memory.

roots must be a slice of references to objects in the heap which are the roots of your object graph. Any object that isn't directly or indirectly referenced by these roots will be freed.

This is the equivalent of heap.mark(roots); heap.sweep();.

pub fn dump(&self) -> String

For debugging: generate a string listing the size and color of each span of memory.

pub fn dump_spans(&self) -> String

For debugging: generate a string listing only the color of each span of memory.

pub fn get_stats(&self) -> HeapStats

Return an object listing the free & total bytes of this heap.